support cuda graph
