+++ python/sglang/srt/model_executor/model_runner.py
@@ -502,10 +502,10 @@ class ModelRunner:
 
     def forward_decode(self, forward_batch: ForwardBatch):
         if self.cuda_graph_runner and self.cuda_graph_runner.can_run(
-            forward_batch.batch_size
-        ):
+            forward_batch.input_ids.numel()
+        ) and forward_batch.forward_mode.is_cuda_graph():
             return self.cuda_graph_runner.replay(forward_batch)
-
+        self.attn_backend.init_forward_metadata(forward_batch)
         return self.model.forward(
             forward_batch.input_ids, forward_batch.positions, forward_batch
         )
