+++ python/sglang/srt/speculative/__init__.py
@@ -1,1 +1,2 @@
 from .eagle_worker import EAGLEWorker
+from . import eagle_utils
